// stock-manager-client/src/app/api/model/produitAjoutDTO.ts
/**
 * API Gestion Produits
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProduitAjoutDTO { 
    nom: string;
    prix: number;
    categorieId: number;
}